Force bg adjustments
Animated background layer now picks the proper canvas color when we
force it to be a certain color.

Added an option to force change the backdrop color in the force bg color
module.

Added an option to only force change the backdrop and canvas colors when
they are default values.